Refactor `session.server.ts` again to fix some bugs

+++ app/session.server.ts
@@ -1,14 +1,16 @@
 import type { SessionIdStorageStrategy } from '@remix-run/node'
-import { createSessionStorage } from '@remix-run/node'
+import { createSessionStorage, redirect } from '@remix-run/node'
 import {
   createSession,
   deleteSession,
+  doesSessionExist,
   getSessionById,
-  updateOrCreateSession,
+  updateSession,
 } from './models/session.server'
 
 const DEFAULT_MAX_AGE = 3600
 const DEFAULT_SECRET = 's3cr3t'
+const INVALID_SESSION_REDIRECT = '/'
 
 function createDatabaseSessionStorage(
   cookie: SessionIdStorageStrategy['cookie']
@@ -16,35 +18,55 @@ function createDatabaseSessionStorage(
   return createSessionStorage({
     cookie,
     async createData(data, expires) {
-      return await createSession(
+      const { id } = await createSession(
         data.userId,
         expires || new Date(Date.now() + DEFAULT_MAX_AGE)
       )
+      return id
     },
     async readData(id) {
-      return await getSessionById(id)
+      const sessionData = await getSessionById(id)
+      return sessionData ?? { invalid: true }
     },
     async updateData(id, data, expires) {
-      if (expires && new Date() > expires) return
-      // TODO: Consider if upserting is necessary
-      await updateOrCreateSession(
+      await updateSession(
         id,
         data.userId,
         expires || new Date(Date.now() + DEFAULT_MAX_AGE)
       )
     },
     async deleteData(id) {
+      if (!(await doesSessionExist(id))) return
       await deleteSession(id)
     },
   })
 }
 
-export const { getSession, commitSession, destroySession } =
-  createDatabaseSessionStorage({
-    name: '__session',
-    path: '/',
-    httpOnly: true,
-    maxAge: DEFAULT_MAX_AGE,
-    sameSite: 'lax',
-    secrets: [process.env.SESSION_SECRET || DEFAULT_SECRET],
-  })
+const {
+  getSession: _getSession,
+  commitSession,
+  destroySession,
+} = createDatabaseSessionStorage({
+  name: '__session',
+  path: '/',
+  httpOnly: true,
+  maxAge: DEFAULT_MAX_AGE,
+  sameSite: 'lax',
+  secrets: [process.env.SESSION_SECRET || DEFAULT_SECRET],
+})
+
+async function getSession(...args: any[]) {
+  const session = await _getSession(...args)
+
+  if (session.has('invalid')) {
+    throw redirect(INVALID_SESSION_REDIRECT, {
+      headers: {
+        'Set-Cookie': await destroySession(session),
+      },
+    })
+  }
+
+  return session
+}
+
+export { getSession, commitSession, destroySession }
